fix thangs and stuff
